Modify legal doc file structure
